+++ src/Client.php
@@ -54,7 +54,7 @@ use Symfony\Component\Uid\Ulid;
 
 class Client
 {
-    private const SDK_VERSION = '1.17.0';
+    private const SDK_VERSION = '1.17.1';
 
     public readonly LoggerInterface $logger;
     public readonly Options $options;