+++ src/Entities/Subscription/SubscriptionNonCatalogProduct.php
@@ -15,15 +15,30 @@ use Paddle\SDK\Entities\Shared\CatalogType;
 use Paddle\SDK\Entities\Shared\CustomData;
 use Paddle\SDK\Entities\Shared\TaxCategory;
 
-class SubscriptionNonCatalogProduct
+/**
+ * @deprecated use Paddle\SDK\Resources\Subscriptions\Operations\Price\SubscriptionNonCatalogProduct instead
+ */
+class SubscriptionNonCatalogProduct implements \JsonSerializable
 {
     public function __construct(
         public string $name,
         public string|null $description,
+        /** @deprecated Not accepted by the API. Ignored during serialization. Will be removed in a future version. */
         public CatalogType|null $type,
         public TaxCategory $taxCategory,
         public string|null $imageUrl,
         public CustomData|null $customData,
     ) {
+    }
+
+    public function jsonSerialize(): array
+    {
+        return [
+            'name' => $this->name,
+            'description' => $this->description,
+            'tax_category' => $this->taxCategory,
+            'image_url' => $this->imageUrl,
+            'custom_data' => $this->customData,
+        ];
     }
 }