+++ src/Resources/Subscriptions/Operations/Charge/SubscriptionChargeNonCatalogPrice.php
@@ -1,0 +1,47 @@
+<?php
+
+declare(strict_types=1);
+
+namespace Paddle\SDK\Resources\Subscriptions\Operations\Charge;
+
+use Paddle\SDK\Entities\Shared\CustomData;
+use Paddle\SDK\Entities\Shared\Money;
+use Paddle\SDK\Entities\Shared\PriceQuantity;
+use Paddle\SDK\Entities\Shared\TaxMode;
+use Paddle\SDK\Entities\Shared\UnitPriceOverride;
+use Paddle\SDK\FiltersUndefined;
+use Paddle\SDK\Undefined;
+
+class SubscriptionChargeNonCatalogPrice implements \JsonSerializable
+{
+    use FiltersUndefined;
+
+    /**
+     * @param array<UnitPriceOverride> $unitPriceOverrides
+     */
+    public function __construct(
+        public string $productId,
+        public string $description,
+        public Money $unitPrice,
+        public string|Undefined|null $name = new Undefined(),
+        public TaxMode|Undefined $taxMode = new Undefined(),
+        public array|Undefined $unitPriceOverrides = new Undefined(),
+        public PriceQuantity|Undefined $quantity = new Undefined(),
+        public CustomData|Undefined|null $customData = new Undefined(),
+    ) {
+    }
+
+    public function jsonSerialize(): array
+    {
+        return $this->filterUndefined([
+            'product_id' => $this->productId,
+            'description' => $this->description,
+            'unit_price' => $this->unitPrice,
+            'name' => $this->name,
+            'tax_mode' => $this->taxMode,
+            'unit_price_overrides' => $this->unitPriceOverrides,
+            'quantity' => $this->quantity,
+            'custom_data' => $this->customData,
+        ]);
+    }
+}